+++ discord/threads.py
@@ -28,7 +28,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Callable, Iterable
 
 from .abc import Messageable, _purge_messages_helper
-from .enums import ChannelType, try_enum
+from .enums import ChannelType, try_enum, ThreadAutoArchiveDuration
 from .errors import ClientException
 from .flags import ChannelFlags
 from .mixins import Hashable
@@ -602,7 +602,7 @@ class Thread(Messageable, Hashable):
         locked: bool = MISSING,
         invitable: bool = MISSING,
         slowmode_delay: int = MISSING,
-        auto_archive_duration: ThreadArchiveDuration = MISSING,
+        auto_archive_duration: ThreadAutoArchiveDuration | ThreadArchiveDuration = MISSING,
         pinned: bool = MISSING,
         applied_tags: list[ForumTag] = MISSING,
         reason: str | None = None,
@@ -632,6 +632,7 @@ class Thread(Messageable, Hashable):
         auto_archive_duration: :class:`int`
             The new duration in minutes before a thread is automatically archived for inactivity.
             Must be one of ``60``, ``1440``, ``4320``, or ``10080``.
+            **ThreadAutoArchiveDuration** enum can be used for better understanding.
         slowmode_delay: :class:`int`
             Specifies the slowmode rate limit for user in this thread, in seconds.
             A value of ``0`` disables slowmode. The maximum value possible is ``21600``.
@@ -662,7 +663,9 @@ class Thread(Messageable, Hashable):
         if archived is not MISSING:
             payload["archived"] = archived
         if auto_archive_duration is not MISSING:
-            payload["auto_archive_duration"] = auto_archive_duration
+            payload["auto_archive_duration"] = auto_archive_duration.value \
+                if isinstance(auto_archive_duration, ThreadAutoArchiveDuration) \
+                else auto_archive_duration
         if locked is not MISSING:
             payload["locked"] = locked
         if invitable is not MISSING: